backend: dedupe tokenizing code in gptj/mpt
